+++ packages/iov-cli/src/cli.ts
@@ -74,6 +74,7 @@ export const main = (originalArgs: string[]): void => {
   console.log(colors.yellow("    - Encoding"));
   console.log(colors.yellow("  * from @iov/faucets"));
   console.log(colors.yellow("    - BovFaucet"));
+  console.log(colors.yellow("    - IovFaucet"));
   console.log(colors.yellow("  * helper functions"));
   console.log(colors.yellow("    - toAscii"));
   console.log(colors.yellow("    - fromHex"));
@@ -109,7 +110,7 @@ export const main = (originalArgs: string[]): void => {
     import { bnsCodec } from '@iov/bns';
     import { Bip39, Ed25519, Ed25519Keypair, Random, Sha256, Sha512 } from '@iov/crypto';
     import { Bech32, Encoding } from '@iov/encoding';
-    import { BovFaucet } from '@iov/faucets';
+    import { BovFaucet, IovFaucet } from '@iov/faucets';
     const { toAscii, fromHex, toHex } = Encoding;
   `;