+++ packages/iov-faucets/types/bovfaucet.d.ts
@@ -2,7 +2,7 @@ import { Address, TokenTicker } from "@iov/bcp-types";
 /**
  * A faucet for the blockchain of value.
  *
- * @deprecated will be removed in 0.9. Migrate to IovFaucet.
+ * @deprecated will be removed in 0.10. Migrate to IovFaucet.
  */
 export declare class BovFaucet {
     private readonly url;